fix file watcher event bug on Linux
the watcher can not normally get events of file changes on linux. it just can get the first two changes.
